clippy: three gates that go red on the current stable, not on this laptop
`cargo clippy --workspace --all-targets -- -D warnings` is a CI job, and
CI runs whatever stable is TODAY. Three findings had accumulated behind
an older local toolchain, so main was red before the release tag was
even cut:

- examples/grounds.rs — `type_complexity` on the row builder's return.
  Named as `GroundRow`, which is also what the printed line IS.
- tests/perf_budgets.rs — `assertions_on_constants`. `cfg!(...)` IS a
  constant, so the debug-build guard is now a branch-and-panic. Behaviour
  is unchanged and checked both ways: debug `--ignored` still fails 10 of
  12 with the same message, release still passes 12 of 12.
- tests/adv_layout.rs — `unnecessary_cast`; `rng.below` already returns
  `usize`.

No behaviour change in any of the three. `cargo test --workspace`: 2603
passed / 0 failed.

+++ tests/adv_layout.rs
@@ -598,7 +598,7 @@ fn content_sized_grid_children_are_solved_big_enough_for_their_own_content() {
     let mut rng = Rng::new(0x0061_1D00);
     let mut load_bearing = 0usize;
     for case in 0..400 {
-        let ncols = 1 + rng.below(4) as usize;
+        let ncols = 1 + rng.below(4);
         let cols: Vec<Track> = (0..ncols)
             .map(|_| match rng.below(4) {
                 0 => Track::Cells(2 + rng.below(12) as i32),
